+++ src/main/kotlin/one/oktw/galaxy/galaxy/planet/data/extensions/Planet.kt
@@ -1,12 +1,12 @@
 package one.oktw.galaxy.galaxy.planet.data.extensions
 
 import one.oktw.galaxy.Main.Companion.galaxyManager
-import one.oktw.galaxy.enums.AccessLevel
-import one.oktw.galaxy.enums.AccessLevel.*
 import one.oktw.galaxy.galaxy.data.extensions.getGroup
 import one.oktw.galaxy.galaxy.enums.Group.VISITOR
 import one.oktw.galaxy.galaxy.planet.PlanetHelper
 import one.oktw.galaxy.galaxy.planet.data.Planet
+import one.oktw.galaxy.galaxy.planet.enums.AccessLevel
+import one.oktw.galaxy.galaxy.planet.enums.AccessLevel.*
 import org.spongepowered.api.entity.living.player.Player
 import org.spongepowered.api.world.World
